refactor(ui): migrate sheet components to components/ui
